feat(DPLAN-11858): move the selected tags logic into the separate component AssignedTags.vue.
